Move images to relative folders
Encapsule relative file to same folder turn the project
more easy to future changes.
